Change GetNumberOfColumns return type to size_t

+++ src/model/idataset_stream.h
@@ -9,7 +9,7 @@ class IDatasetStream {
 public:
     virtual std::vector<std::string> GetNextRow() = 0;
     [[nodiscard]] virtual bool HasNextRow() const = 0;
-    [[nodiscard]] virtual int GetNumberOfColumns() const = 0;
+    [[nodiscard]] virtual size_t GetNumberOfColumns() const = 0;
     [[nodiscard]] virtual std::string GetColumnName(int index) const = 0;
     [[nodiscard]] virtual std::string GetRelationName() const = 0;
     virtual void Reset() = 0;
